requirements (1).txt
web3==6.15.1
python-telegram-bot==21.3
requests==2.31.0
websockets==12.0
aiohttp==3.9.5
